Added Audio Recorder Fragment to Phone App
Both on the backend and frontend, meaning you can click on it. It saves to the device and when you click start recording it will print out the file directory. Lastly it still needs to bind to the loomo, although it will record though your phone and save on your phone.